pyproject.toml
[project]
name = "zfsnappr"
version = "1.2.6"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "pschlo", email = "70105856+pschlo@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dateutil>=2.9.0.post0",
]

[project.scripts]
zfsnappr = "zfsnappr.entrypoint:cli"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"
